Option to define types once.
Previously cxxbridge emitted type definitions twice, once in the .cc and
once in the .h file.

This is undesirable when Arbitrarily Complex Build systems end up including
the .h in the .cc.

One option considered was to wrap the type definitions in a
  #ifndef STUFF_DEFINED
  #define STUFF_DEFINED
  ...
  #endif
section, but the 'STUFF_DEFINED' preprocessor symbol would need to have a
unique name for every such section to avoid potential conflicts in codebases
with multiple cxxbridge invocations and potentially duplicate filenames
(foo/path.h and bar/path.h).

In any case it seems more normal to define things *once* in C++, and require
the .cc file to #include the header file. This change therefore makes that
option available. Specifically, we rely on an existing option to provide
an include path to cxxbridge, and just add another option to cease to
redefine types.