Allow impl UniquePtrTarget for T;
cxx allows UniquePtr<T> only if T implements a trait,
UniquePtrTarget. This trait implementation is automatically generated
for type T only if:

* UniquePtr<T> is used in a struct or a function parameter/return
  type; AND
* the type T is a C++ type.

Until now, it was not possible to generate UniquePtrTarget
implementations for type aliases where a cxx type refers to an
externally defined type (e.g. from bindgen).

There were also legitimate uses for UniquePtr<T> even if none
of the structs/functions within the cxx::bridge actually used it;
for example storing a UniquePtr<T> in some other struct outside
the cxx::bridge mod.

This change allows any cxx::bridge block to contain
  impl UniquePtrTarget for T {}
which will force cxx to generate the UniquePtrTarget trait (and
associated C++ code) for that type.

It's not desirable to do this for _every_ type since it does
generate additional C++ code as well as Rust code, and thus in some
toolchains may cause binary bloat.

Future changes could allow similar syntax for types which should
be containable in a C++ vector, e.g.:
  impl VectorElement for T {}
but that's not yet done here.