Tweak wording of the triviality requirement
It isn't actually a requirement that the type have no destructor as
brought up by the previous paragraph. Rather, it just needs to be safe
for values of that type to be moved via Rust's memcpy-based move
semantics.

The distinction is relevant e.g. for `struct S { s: String }` which
would have a C++ destructor and nontrivial move constructor but still be
safely memcpy movable.