Output fully-qualified names.
This change:
* Outputs fully-qualified C++ names wherever possible
  or appropriate;
* Supports outputting content belonging to different
  namespaces.

This change should have no functional effect, because
at the moment the only way to set up a namespace is still
to configure the overall [cxx::bridge] with a namespace
attribute. However, in theory, it should allow finer-grained
namespaces to be supported in subsequent commits. In theory,
it should be possible for items in one namespace to correctly
reference another thanks to the fully qualified names.