Check std::is_enum on extern enum
Without this, something that is not an enum could have been able to pass
our assertions, resulting in unpredictable behavior.

    // rust
    enum FakeEnum { K = 1usize }
    extern "C++" { type FakeEnum; }

    // c++
    struct FakeEnum {
      static const int K = 1;
      const char *x;
    };

There may be valid reasons to want something of this sort to work, but
forbid for now and someone will need to bring forward a concrete use
case for consideration.