Add UniquePtr::as_mut_ptr
cxx accepts functions that take *mut T as an argument.
Such functions are of course used under odd niche circumstances;
it would be safer and better to take a &mut Pin<T>.

But in cases where such a function does exist, we'd most commonly
want to provide it a mutable pointer to a T which is safely
and uniquely owned by Rust, within a UniquePtr.

It was previously quite hard to do this:

    let mut a = /* make UniquePtr to a thing */
    unsafe { ffi::TakeA(std::pin::Pin::<&mut ffi::A>::into_inner_unchecked(a.pin_mut())) };

With this extra API, it becomes much more ergonomic:

    let mut a = /* make UniquePtr to a thing */
    unsafe { ffi::TakeA(a.as_mut_ptr()) }