bazel: don't use cargo in $PATH, make the build hermetic.
Background:
tools/bazel/vendor is just running 'cargo', hoping to find it somewhere
in $PATH.

Problem:
rules_rust try to be hermetic, and not use the cargo of the system, but
the cargo of the toolchain installed.
This means that if the cargo on the system is different from the cargo
used by cxx.rs, or the system does not have a cargo at all, bad things
will happen when using cxx.rs.

In this change:
use the recommended way to find/install the tool. Thanks to Andre in the
bazel rust slack channel for pointers and help.