Add CXX_RS_EXPORT, CXX_CPP_EXPORT macros to cxx.cc.
In some non-Cargo build configurations, the Rust and C++ side of cxx
bindings may end up in different binaries. In some environments,
this requires symbols to be explicitly exported using
__declspec(dllexport) or similar directives. A previous change added
the --cxx-impl-annotations switch on the cxx-gen command line tool,
providing the opportunity to specify such directives for dynamically
generated parts of the cxx bindings.

It turns out to be similarly useful to export symbols from cxx.cc.
This change adds a pair of preprocessor symbols which folks can define
if they wish to export these symbols too.