gen: Fix panic in clap when handling `--include`
For some reason I don't yet fully understand, the build
of rpm-ostree started failing with a panic in `cxxbridge`:
https://github.com/coreos/rpm-ostree/pull/3763#issuecomment-1154193336

What I haven't worked out is what changed to break this; my best
guess is something like compiler type inference?  AFAICS, the
code here hasn't changed in a long time, and neither has the lockfiles.

(Hmm...unless we're not using lockfiles for this...ah yes, that would
 explain it.  And https://crates.io/crates/clap/3.2.1 was just published.
 Will dig into this more)

Anyways, this is not an elegant fix; we just (correctly) ask clap
to give us `OsString` and not `String`.  The real fix is clearly
to stop using all the deprecated clap APIs, but I'm trying to make
the least invasive change for the moment.