tests/ffi: "fix" the error declaring a struct for Buffer
I'm not sure what's happening here, I had the doubt that the
tuple like struct was not supported, so I moved to a plain
struct with named fields and the error disappeared.

I think that this is not a solution, but only that I have
hidden again some issue in wrapping Trivial structs.