Opt out -Zrustdoc-scrape-examples on docs.rs
I'd like a chance to audit all the code that rustdoc is inserting into
the docs. Currently I am skeptical that showing cxx's internal usages of
APIs is a net benefit to the public documentation. I am also skeptical
that quite so many examples are needed, and that they should be featured
so prominently in comparison to handwritten docs. Lastly I wish there
were a way to turn this behavior off on a more granular basis.