Fix: Prevent crash on OOM in error copy
When the error message is copied, the allocation may fail. This would
terminate the process, since it's called from a `noexcept` function.

Make the allocation `nothrow` and handle `nullptr` in `what()` method
appropriately, returning a generic message saying that the message
cannot be allocated.

This also prevents a crash if `what()` is called on a moved object
(which is technically UB, but still better not to crash).