chore: clippy let-underscore-untyped
Clippy keeps complaining about
  https://rust-lang.github.io/rust-clippy/master/index.html#let_underscore_untyped

Seems like a noop, esp considering that @dtolnay has filed a bug in
  https://github.com/rust-lang/rust-clippy/issues/10411

One issue though is that MSRV in .clippy.toml is not matching the one in Cargo.toml:
* .clippy.toml:  msrv = "1.48.0"
* Cargo.toml: rust-version = "1.60"

Which one is the accurate one?  I would assume its the Cargo.toml, in
which case it might be better to apply this change?