Add template deduction guide for Slice constructor
PR #1367 added an explicit Slice constructor which allows constructing
from any C++ container.

This adds a template deduction guide for that constructor which allows
the Slice type to be inferred from the container's contained type,
taking into account the const-ness of the type pointed to by data().

This applies only to C++17 and newer.

Test: cargo test
Test: cargo test -F c++14
Test: cargo test -F c++17
Test: cargo test -F c++20