macro: ensure that cfg attrs are set for type alias verify generation
```rust
mod ffi {
  unsafe extern "C++" {
    #[cfg(enabled)]
    type A = crate::A;
  }
}
```

When using the bridge above we need to ensure that all generated code
also has the cfg attribute set. Before this change the following error
would occur `cannot find type A in this scope`. This appears to be
due to the generated verify block below that was not copying the
attributes from the original type.

```rust
const _: fn() = ::cxx::private::verify_external_type::<A, ...>;
````

After this change the attributes are copied and therefore when the type
is not compiled the verify block is also not compiled.