[workerd-cxx] eliminate source copying (#28)
* [workerd-cxx] eliminate source copying

The original cxx crate uses symlinks to share code between its
sub-crates. I assume this is because of cargo limitation that
forces you to share such shared code as an independent crate that
anyone can use.

Symlinks don't work well on windows, so we got rid of them by copying
the code during compilation. This has the unfortunate effect of
rust-analyzer (and other tools) not working properly.

We currently consume workerd-cxx only through bazel and we do not
have such limitations.

This change eliminates the copy and adjusts visibilities (and some
minor detail of implementation) for things to compile.

I hope that usability gains will overweight the additional merge
difficulties.

* clippy