[workerd-cxx] remove + Send
KJ event loop runs only on a single thread, so it is not legal to pass
futures/promises to a different one. Removing +Send achieves this.

Green threads while technically pass event loops between threads,
they are creating a mirage of staying within the same thread. So
we can use existing type system protection.

This doesn't take opportinity to simplify the rest of the code
(it follows).

With this change the following code now fails appropriately:

	let x = ffi::new_ready_promise_void();
	std::thread::spawn(async || {
		let _ = x.await;
	});