Improving support for cross-bridge type aliases of Rust types.
This commit improves docs (to explain that the type alias has to go into
a `extern "C++"` section, and to point out the need to use
`#[derive(ExternType)]`).

This commit also fixes errors that would be reported if a type alias for
a Rust type would be used as a `&mut T` parameter.  The fixes are to
ensure that the Rust type is recognized as Rust-movable.  More details
can be found in the PR discussion.