Don't track bridge files in `OUT_DIR`
Files in the `OUT_DIR` are likely generated by the very same build script. As such, they likely have a modification time that is after the build script's start time. This causes cargo to assume that the file has changed since the last run, thereby considering the build script dirty on every build. The result is that the build script and everything above in the dependency is recompiled on every build.

This commit avoids that problem by emitting the `rerun-if-changed` command only for files that are not in the `OUT_DIR`.