Replace `fn local_type` in `generics.rs` with more granular functions.
This commit replaces `fn local_type(ty: &Type) -> &NamedType` with
two more granular functions:

* `fn get_impl_generics(ty: &Type, types: &Types) -> &Lifetimes`
* `fn format_for_prevent_unwind_label(ty: &Type) -> TokenStream`

This refactoring is desirable, because it makes it easier to expand
these functions to support more than just `Type::Ident` types.  For
example `Type::RustVec(ty1)` may not have a `NamedType` of its own, but
it can still be translated into `&Lifetimes` and/or `TokenStream`.