Early error detaction when `cxx_name` is a reserved C++ keyword.
Before this commit, `cxx` wouldn't detect the problem, and later the C++
compilation would fail.  After this commit, the error is detected early.

This commit seems desirable on its own, but it also helps to build
scaffolding for enabling additional checks in follow-up commits (e.g.
allowing `operator==` for function names, but not for other
`[cxx::bridge]` API names).